requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
psycopg2==2.9.9
beautifulsoup4==4.12.3
pandas==2.2.1
openpyxl==3.1.2


